% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitFacet.R
\name{splitFacet}
\alias{splitFacet}
\title{Split multi-facet plot into individual plots}
\usage{
splitFacet(x)
}
\arguments{
\item{x}{A multi-facet plot made from ggplot or other R packages}

\item{...}{Other arguments passed on to methods. Not currently used.}
}
\description{
This function allows you to split a multi-facet plot into individual plots.
}
\details{
The function is built along these steps :
(1) it goes through the structure of the object to get the names of the variables used for faceting.
(2) it overwrites the facet element of the plot object with the one from the empty ggplot object (so if we print it at this stage facets are gone).
(3) it extracts the data and split it along the variables identified in the 1st step.
(4) it overwrites the original data with each subset and store all outputs in a list.
}
\examples{
# Generate some sample data, then create a multi-facet plot
df <- data.frame(x=seq(1,24,1), y=seq(1,24,1), z=rep(seq(1,12),each=2))
myplot <- ggplot(df,aes(x=x, y=y))+geom_point()+facet_wrap(~z)
myplot

# Use the 'splitFacet()' function to split into individual plots
new_plots <- splitFacet(myplot)

length(new_plots)
new_plots[[3]]

}
\keyword{plot}
